@@ -1,17 +1,13 @@
 import vueModaltor from "./components/vue-modaltor.vue";
 const modalTor = {
-    install(Vue, options) {
-        let component = vueModaltor;
+    install(app, options = {}) {
+        const component = vueModaltor;
         for (let k in options) {
-            // eslint-disable-next-line no-prototype-builtins
-            if (component.props.hasOwnProperty(k)) {
+            if (Object.prototype.hasOwnProperty.call(component.props, k)) {
                 component.props[k].default = options[k];
             }
         }
-        Vue.component("vue-modaltor", component);
+        app.component("vue-modaltor", component);
     }
 };
-if (typeof window !== "undefined" && window.Vue) {
-    window.Vue.use(modalTor);
-}
 export default modalTor;
